+++ CHANGELOG.md
@@ -4,7 +4,8 @@ based on [Semantic Versioning](http://semver.org/)
 
 *0.2.1* 2015-03-09
 
-+ move the accident method from RedCar in the crash example to MovablePiece 
++ move the accident method from RedCar in the crash example to MovablePiece
++ rename the accident method to hit
 + remove the FallingPiece class in pieces
 + remove the FallingPieces class in pieces
 + replace the move method with move([Direction direction]) of MovablePiece in pieces
@@ -13,6 +14,7 @@ based on [Semantic Versioning](http://semver.org/)
 + add 'black' and 'white' to color list in util
 + rename randomNumber to randomSignNum in util
 + add more random tests
++ add dash example
 
 *0.2.1* 2015-03-08